firebase/functions/requirements.txt
# Firebase Functions Framework
firebase_functions~=0.1.0
firebase-admin>=6.4.0

# AI and Language Processing
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
openai==1.54.5

# HTTP and Web Scraping
httpx==0.27.2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image Processing
Pillow>=10.0.0

# Environment and Configuration
python-dotenv>=1.0.0